Detonated bombs are removed from the field

+++ gui/GameArena.h
@@ -60,7 +60,7 @@ public:
     void addBomb(int player, int x, int y, int bombId);
     void addFlame(Flame& flame);
     void removeFlame(int flameId);
-    void removeBomb(int bombId);
+    void removeBomb(qint16 bombId);
 	void updateMap(QByteArray& updateBlock);
     int getNbPlayers() const;
     const Map *getMap();
